moved config to internal
